+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Unreleased
+- implement `IntoIterator` for `&OneOrMany<T>` and `&mut OneOrMany<T>`
+- add `check` function for verifying an IRI's authority
+- add `BaseExt::check_authority` for verifying an IRI's authority against an object's ID
+- add back checked `Base::id`, `Activity::actor`, `Activity::object`, `Actor::inbox`,
+  `Actor::outbox`, `Actor::following`, `Actor::followers`, `Actor::liked`, `Actor::streams`,
+  `Actor::endpoints`
 
 # 0.7.0-alpha.14
 - switch to iri-string from url